+++ mark1/android/app/src/main/kotlin/com/lukecreated/mymy/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.lukecreated.mymy
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()